@@ -3,40 +3,43 @@
   "version": "1.6.0",
   "private": true,
   "dependencies": {
+    "@babel/runtime": "^7.17.7",
+    "@emotion/react": "^11.8.2",
+    "@emotion/styled": "^11.8.1",
     "@material-ui/core": "^4.12.3",
-    "@material-ui/data-grid": "^4.0.0-alpha.35",
+    "@material-ui/data-grid": "^4.0.0-alpha.37",
     "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.57",
-    "@testing-library/jest-dom": "^5.11.1",
-    "@testing-library/react": "^10.4.7",
-    "@testing-library/user-event": "^12.0.11",
+    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.4",
+    "@testing-library/user-event": "^12.8.3",
     "file-saver": "^2.0.5",
-    "husky": "^7.0.2",
+    "husky": "^7.0.4",
     "jszip": "^3.7.1",
-    "konva": "^7.0.3",
+    "konva": "^7.2.5",
     "lint-staged": "^10.5.4",
-    "material-ui-popup-state": "^1.8.3",
+    "material-ui-popup-state": "^2.0.0",
     "notistack": "^1.0.10",
-    "prettier": "^2.2.1",
-    "qs": "^6.9.4",
-    "react": "^16.13.1",
-    "react-debounce-input": "^3.2.2",
-    "react-dom": "^16.13.1",
-    "react-hotkeys-hook": "^2.4.0",
-    "react-joyride": "^2.3.1",
-    "react-konva": "^16.13.0-3",
-    "react-material-ui-form-validator": "^2.1.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "socket.io-client": "^2.4.0",
-    "typescript": "^4.3.4",
-    "use-image": "^1.0.6"
+    "prettier": "^2.6.0",
+    "qs": "^6.10.3",
+    "react": "^17.0.2",
+    "react-debounce-input": "^3.2.5",
+    "react-dom": "^17.0.2",
+    "react-hotkeys-hook": "^3.4.4",
+    "react-joyride": "^2.3.2",
+    "react-konva": "^17.0.2-5",
+    "react-material-ui-form-validator": "^3.0.0",
+    "react-router-dom": "^6.2.2",
+    "react-scripts": "^5.0.0",
+    "socket.io-client": "^4.4.1",
+    "typescript": "^4.6.2",
+    "use-image": "^1.0.10"
   },
   "scripts": {
     "start": "chmod +x ./env.sh && sh ./env.sh && cp env-config.js ./public/ && react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "test:cy": "cypress run-ct"
+    "test:cy": "cypress run-ct --browser chrome"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -58,19 +61,18 @@
     ]
   },
   "devDependencies": {
-    "@cypress/react": "^5.9.1",
-    "@cypress/webpack-dev-server": "^1.4.0",
-    "@types/file-saver": "^2.0.4",
-    "@types/jest": "^26.0.20",
-    "@types/lodash": "^4.14.170",
-    "@types/node": "^15.12.4",
-    "@types/qs": "^6.9.5",
-    "@types/react": "^16.9.43",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-material-ui-form-validator": "^2.1.0",
-    "@types/react-router-dom": "^5.1.5",
-    "@types/socket.io-client": "^1.4.33",
-    "@visual-regression-tracker/agent-cypress": "^5.0.0",
-    "cypress": "^7.6.0"
+    "@cypress/react": "^5.12.4",
+    "@cypress/webpack-dev-server": "^1.8.3",
+    "@types/file-saver": "^2.0.5",
+    "@types/jest": "^27.4.1",
+    "@types/lodash": "^4.14.180",
+    "@types/node": "^17.0.21",
+    "@types/qs": "^6.9.7",
+    "@types/react": "^17.0.40",
+    "@types/react-dom": "^17.0.13",
+    "@types/react-material-ui-form-validator": "^2.1.1",
+    "@types/react-router-dom": "^5.3.3",
+    "@visual-regression-tracker/agent-cypress": "^5.2.0",
+    "cypress": "^9.5.2"
   }
 }
