@@ -7,14 +7,16 @@ import {
   TestRunProvider,
   HelpProvider,
 } from "../contexts";
-import { MemoryRouter, Route } from "react-router-dom";
+import { MemoryRouter, Route, Routes } from "react-router-dom";
 import { MemoryRouterProps } from "react-router";
 import { SnackbarProvider } from "notistack";
 
 export const mountVrtComponent = ({
   component,
-  memoryRouterProps,
-  path,
+  memoryRouterProps = {
+    initialEntries: ["/"],
+  },
+  path = "/",
 }: {
   component: React.ReactElement;
   memoryRouterProps?: MemoryRouterProps;
@@ -22,21 +24,26 @@ export const mountVrtComponent = ({
 }) =>
   mount(
     <MemoryRouter {...memoryRouterProps}>
-      <Route path={path}>
-        <SnackbarProvider>
-          <UserProvider>
-            <ProjectProvider>
-              <BuildProvider>
-                <HelpProvider>
-                  <TestRunProvider>{component}</TestRunProvider>
-                </HelpProvider>
-              </BuildProvider>
-            </ProjectProvider>
-          </UserProvider>
-        </SnackbarProvider>
-      </Route>
+      <Routes>
+        <Route
+          path={path}
+          element={
+            <SnackbarProvider>
+              <UserProvider>
+                <ProjectProvider>
+                  <BuildProvider>
+                    <HelpProvider>
+                      <TestRunProvider>{component}</TestRunProvider>
+                    </HelpProvider>
+                  </BuildProvider>
+                </ProjectProvider>
+              </UserProvider>
+            </SnackbarProvider>
+          }
+        />
+      </Routes>
     </MemoryRouter>,
     {
-      stylesheets: ["/src/index.css"],
+      cssFiles: ["src/index.css"],
     }
   );
